Fix growfs service
i had an issue with the `ConditionFirstBoot=true`, so i created my own
way to check if the first boot happened or not by creating a file under
the `/etc/seabot-firstboot`, this file cretaed by the `seabot-growfs.sh`
script [it has an an easter egg, please dont remove it], also the
service checks teh existance of that file to know if it should be run or
not.

update the seabot-fetch script to show the current disk space and usage.